// web/projects/portal/src/app/common/data/file-types.ts
/*
 * inetsoft-web - StyleBI is a business intelligence web application.
 * Copyright © 2024 InetSoft Technology (info@inetsoft.com)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
export enum FileTypes {
   TEMPLATE = 1,
   REPORT = 2,
   CSV = 3,
   RTF = 4,
   RTF_LAYOUT = 6,
   HTML = 5,
   HTML_BUNDLE = 7,
   SVG = 8,
   PDF = 9,
   HTML_NO_PAGINATION = 10,
   TEXT = 11,
   HTML_BUNDLE_NO_PAGINATION = 12,
   POWERPOINT = 13,
   POSTSCRIPT = 14,
   XML = 15,
   HTML_NO_PAGINATION_EMAIL = 16,
   GENERATED_REPORT = 1002,
   EXCEL = 1003,
   EXCEL_SHEET = 1004,
   EXCEL_DATA = 1005,
   BURST_REPORT = 2001,
   USER_DEFINED = 5000
}